+++ library/config/dataclasses/sd_peft.py
@@ -3,7 +3,7 @@ from .training import TrainingConfig
 from .optimizer import OptimizerConfig
 from .dataset import DatasetConfig
 from .buckets import BucketsConfig
-from .sd_models import SDModelsConfig
+from .model import ModelConfig
 from .network import NetworkConfig
 from .saving import SavingConfig
 from .huggingface import HuggingFaceConfig
@@ -18,12 +18,12 @@ from .metadata import MetadataConfig
 
 
 @dataclass
-class TrainNetworkConfig:
+class SDPeftConfig:
     training: TrainingConfig = field(default_factory=TrainingConfig)
     optimizer: OptimizerConfig = field(default_factory=OptimizerConfig)
     dataset: DatasetConfig = field(default_factory=DatasetConfig)
     buckets: BucketsConfig = field(default_factory=BucketsConfig)
-    sd_models: SDModelsConfig = field(default_factory=SDModelsConfig)
+    sd_models: ModelConfig = field(default_factory=ModelConfig)
     network: NetworkConfig = field(default_factory=NetworkConfig)
     saving: SavingConfig = field(default_factory=SavingConfig)
     huggingface: HuggingFaceConfig = field(default_factory=HuggingFaceConfig)