+++ library/config/dataclasses/sd_textual_inversion.py
@@ -4,8 +4,7 @@ from .training import TrainingConfig
 from .optimizer import OptimizerConfig
 from .dataset import DatasetConfig
 from .buckets import BucketsConfig
-from .sd_models import SDModelsConfig
-from .network import NetworkConfig
+from .model import ModelConfig
 from .saving import SavingConfig
 from .huggingface import HuggingFaceConfig
 from .logging import LoggingConfig
@@ -33,7 +32,7 @@ class TextualInversionConfig:
     optimizer: OptimizerConfig = field(default_factory=OptimizerConfig)
     dataset: DatasetConfig = field(default_factory=DatasetConfig)
     buckets: BucketsConfig = field(default_factory=BucketsConfig)
-    sd_models: SDModelsConfig = field(default_factory=SDModelsConfig)
+    sd_models: ModelConfig = field(default_factory=ModelConfig)
     saving: SavingConfig = field(default_factory=SavingConfig)
     huggingface: HuggingFaceConfig = field(default_factory=HuggingFaceConfig)
     logging: LoggingConfig = field(default_factory=LoggingConfig)