+++ library/config/dataclasses/sdxl_textual_inversion.py
@@ -4,8 +4,8 @@ from .training import TrainingConfig
 from .optimizer import OptimizerConfig
 from .dataset import DatasetConfig
 from .buckets import BucketsConfig
-from .sd_models import SDModelsConfig
-from .sdxl_training import SDXLTrainingConfig
+from .model import ModelConfig
+from .sdxl import SDXLConfig
 from .saving import SavingConfig
 from .huggingface import HuggingFaceConfig
 from .logging import LoggingConfig
@@ -40,8 +40,8 @@ class SDXLTextualInversionConfig:
     optimizer: OptimizerConfig = field(default_factory=OptimizerConfig)
     dataset: DatasetConfig = field(default_factory=DatasetConfig)
     buckets: BucketsConfig = field(default_factory=BucketsConfig)
-    sd_models: SDModelsConfig = field(default_factory=SDModelsConfig)
-    sdxl_training: SDXLTrainingConfig = field(default_factory=SDXLTrainingConfig)
+    sd_models: ModelConfig = field(default_factory=ModelConfig)
+    sdxl: SDXLConfig = field(default_factory=SDXLConfig)
     saving: SavingConfig = field(default_factory=SavingConfig)
     huggingface: HuggingFaceConfig = field(default_factory=HuggingFaceConfig)
     logging: LoggingConfig = field(default_factory=LoggingConfig)